requirements.txt
Django==2.0.4
django-email-from-template==2.3.1
django-yadt==2.2.4
psycopg2==2.7.3.1
https://github.com/lamby/django-setuptest/zipball/django-1.9-support#egg=django-setuptest
django-staticfiles-dotd==2.0.0
pyScss==1.3.5
django-formtools==2.1
django-switch-templatetag==2.0.4
jsonfield==2.0.2
django-enumfields==0.10.0
django-auto-one-to-one==2.2.1
https://github.com/lamby/django-two-factor-auth/zipball/anonymous-django-2#egg=django-two-factor-auth
gunicorn==19.7.1
python-memcached==1.58
django-storages==1.6.5
boto==2.48.0
watchtower==0.5.2
envparse==0.2.0
https://github.com/lamby/python-redis-rate-limit/zipball/master#egg=python-redis-rate-limit
redis==2.10.6
django_keyerror==3.5.2
GeoIP2==2.6.0
olefile==0.44
Pillow==4.3.0
django-markdown-deux==1.0.5
